profile page cleanup + make profile module
