+++ frontend/src/components/CreateEvent.js
@@ -20,6 +20,7 @@ function mapDispatchToProps(dispatch) {
 
 const categories = ["Social Gathering", "Volunteer Event", "Student Organization Event"];
 const url="https://maps.googleapis.com/maps/api/js?key=" + process.env.REACT_APP_API_KEY + "&libraries=places"; 
+
 class Events extends Component {
   constructor(props) {
     super(props);
