+++ frontend/src/components/UserEvents.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+class UserEvents extends React.Component {
+  constructor(props) {
+    super(props);
+  }
+}
